[Refactor] 중복 사용되는 utils 분리
