+++ js/detail.js
@@ -1,3 +1,5 @@
+import { showToast, showLoadingAnimation, hideLoadingAnimation } from './utils.js';
+
 // firebase 기본 설정
 // import firebaseConfig from '../config.js';
 import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.6.4/firebase-app.js';
@@ -18,32 +20,6 @@ const firebaseConfig = {
 const firebaseApp = initializeApp(firebaseConfig);
 const database = getDatabase(firebaseApp);
 const storage = getStorage(firebaseApp);
-
-
-// toast message
-const toast = document.querySelector('.toast-wrap');
-
-function showToast(message) {
-    const toastMessage = document.querySelector('.toast-message');
-    toastMessage.textContent = message;
-
-    toast.style.display = 'block';
-    setTimeout(() => {
-        toast.style.display = 'none';
-    }, 2000); 
-}
-
-
-// 로딩 애니메이션
-const loadingAnimation = document.querySelector('.loading-animation');
-
-function showLoadingAnimation() {
-    loadingAnimation.style.display = 'block';
-}
-
-function hideLoadingAnimation() {
-    loadingAnimation.style.display = 'none';
-}
 
 
 document.addEventListener('DOMContentLoaded', () => {
